insectopedia-ask-me-main/backend/requirements.txt
# RAG System Backend Dependencies

# Web Framework
flask==2.3.3
flask-cors==4.0.0

# Data Processing
pandas==2.1.1
numpy==1.24.3

# AI/ML Libraries
google-generativeai==0.3.2
sentence-transformers==2.2.2
faiss-cpu==1.7.4
scikit-learn==1.3.0

# Utilities
python-dotenv==1.0.0
werkzeug==2.3.7

# Optional: For advanced text processing
nltk==3.8.1
spacy==3.6.1